renamed modules and endpoints: diff->list, traceDiff->diff
