+++ report-jtreg-diff/src/main/java/io/jenkins/plugins/report/jtreg/main/diff/DiffInfo.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.report.jtreg.main.tracediff;
+package io.jenkins.plugins.report.jtreg.main.diff;
 
 import io.jenkins.plugins.report.jtreg.CommonOptions;
 import io.jenkins.plugins.report.jtreg.formatters.BasicFormatter;
