+++ report-jtreg-list/src/main/java/io/jenkins/plugins/report/jtreg/main/list/cmdline/Options.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package io.jenkins.plugins.report.jtreg.main.diff.cmdline;
+package io.jenkins.plugins.report.jtreg.main.list.cmdline;
 
 import io.jenkins.plugins.report.jtreg.formatters.ColorFormatter;
 import io.jenkins.plugins.report.jtreg.formatters.HtmlFormatter;
