+++ report-jtreg-service/pom.xml
@@ -15,7 +15,7 @@
     <dependencies>
         <dependency>
             <groupId>io.jenkins.plugins</groupId>
-            <artifactId>report-jtreg-diff</artifactId>
+            <artifactId>report-jtreg-list</artifactId>
             <version>${revision}${changelist}</version>
         </dependency>
         <dependency>
@@ -25,7 +25,7 @@
         </dependency>
         <dependency>
             <groupId>io.jenkins.plugins</groupId>
-            <artifactId>report-jtreg-tracediff</artifactId>
+            <artifactId>report-jtreg-diff</artifactId>
             <version>${revision}${changelist}</version>
             <scope>compile</scope>
         </dependency>
